crud funcional usuario

+++ demo/src/main/java/com/hotel/demo/models/Dao/IUsuarioDao.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.hotel.demo.models.entity.Usuario;
 
 @Repository
-public interface IUsuarioDao extends JpaRepository<Usuario, Integer>{
+public interface IUsuarioDao extends JpaRepository<Usuario, Long>{
